+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v2.0.1
+
+- (#20) Fixes inventory failure when platform is not defined for a device
+## v2.0.0
+
+- (#18) Migrates functions to new NTC netutils library, which is removing methods previously available:
+  - compliance
+  - make_folder
+  - hostname_resolves
+  - test_tcp_port
+  - is_ip
+
 ## v0.1.0 - 2020-12-27
 
 Initial release